Flytte/samle logikk utledning reguleringstype
