--- domain/src/main/kotlin/no/nav/su/se/bakover/domain/regulering/UtledReguleringstypeOgNyttFradrag.kt
@@ -1,160 +1,0 @@
-package no.nav.su.se.bakover.domain.regulering
-
-import arrow.core.Either
-import arrow.core.left
-import arrow.core.right
-import no.nav.su.se.bakover.common.domain.extensions.filterLefts
-import no.nav.su.se.bakover.common.domain.extensions.filterRights
-import vilkår.inntekt.domain.grunnlag.FradragTilhører
-import vilkår.inntekt.domain.grunnlag.Fradragsgrunnlag
-import vilkår.inntekt.domain.grunnlag.Fradragstype
-import java.math.BigDecimal
-import java.time.Month
-
-/**
- * Utleder reguleringstype (automatisk/manuell) basert på fradrag brukt fra vedtaksdata og oppdaterer
- * dem med regulerte beløper hentet fra ekstern kilde.
- *
- * @param fradrag Liste med fradragsgrunnlag fra vedtaksdata (krever at det er vedtaksdata fra og med 1. mai)
- * @param eksterntRegulerteBeløp inneholder beløp før og etter regulering
- *          for både bruker og ektefelle/samboer (EPS)
- *
- * @return Par som inneholder:
- *         - Først: Reguleringstype (AUTOMATISK hvis alle fradrag kan behandles automatisk,
- *           eller MANUELL med et sett av årsaker hvis manuell behandling er nødvendig)
- *         - Andre: Liste med fradragsgrunnlag, oppdatert med nye beløp der ekstern regulering
- *           var tilgjengelig og gyldig, sortert etter periode
- *          Eller feiltype [ÅrsakRevurdering] hvis det er differanse mellom vårt og eksternt beløp
- *
- * ## Reguleringslogikk:
- * - **Automatisk regulering** skjer når:
- *   - Eksternt regulerte beløp stemmer overens med våre beløp før regulering
- *   - Alle fradrag har tilhørende eksterne data
- *
- * - **Manuell regulering** er påkrevd når:
- *   - En fradragstype ikke kan justeres automatisk fordi vi ikke har en automatisk kilde/integrasjon (f.eks. Kvalifiseringsstønad)
- *   - Det er en differanse i beløp før regulering
- *   - Eksterne reguleringsdata mangler for et fradrag
- *
- * @see Reguleringstype
- * @see ÅrsakTilManuellRegulering
- * @see måRevurderePåGrunnAvDifferanseMedEksterneBeløp
- */
-fun utledReguleringstypeOgOppdaterFradrag(
-    fradrag: List<Fradragsgrunnlag>,
-    eksterntRegulerteBeløp: EksterntRegulerteBeløp,
-): Either<ÅrsakRevurdering, Pair<Reguleringstype, List<Fradragsgrunnlag>>> {
-    if (fradrag.any { it.periode.fraOgMed.month < Month.MAY }) {
-        throw IllegalArgumentException("Regulering skal ikke kjøres med vedtaksdata før mai året reguleringen kjører i")
-    }
-    val utledetReguleringstypePerFradrag = fradrag.filter { it.periode.fraOgMed.month == Month.MAY }.map {
-        utledPerFradragstypeOgTilhørende(it, eksterntRegulerteBeløp)
-    }
-    if (utledetReguleringstypePerFradrag.any { it.isLeft() }) {
-        return ÅrsakRevurdering(
-            årsak = ÅrsakRevurdering.Årsak.DIFFERANSE_MED_EKSTERNE_BELØP,
-            diffBeløp = utledetReguleringstypePerFradrag.filterLefts(),
-        ).left()
-    }
-    return utledetReguleringstypePerFradrag.filterRights().let {
-        val utledetReguleringstypeForAlleFradrag = it.map { it.first }
-        val reguleringstype =
-            if (utledetReguleringstypeForAlleFradrag.any { it is Reguleringstype.MANUELL }) {
-                Reguleringstype.MANUELL(
-                    problemer = (
-                        utledetReguleringstypeForAlleFradrag
-                            .filterIsInstance<Reguleringstype.MANUELL>()
-                        )
-                        .flatMap { it.problemer }.toSet(),
-                )
-            } else {
-                if (fradrag.any { it.periode.fraOgMed.month > Month.MAY }) {
-                    Reguleringstype.MANUELL(
-                        problemer = setOf(ÅrsakTilManuellRegulering.EtAutomatiskFradragHarFremtidigPeriode()),
-                    )
-                } else {
-                    Reguleringstype.AUTOMATISK
-                }
-            }
-
-        val oppdaterteFradrag = it.map { it.second }
-            .sortedWith(compareBy<Fradragsgrunnlag> { it.periode.fraOgMed }.thenBy { it.periode.tilOgMed })
-        (reguleringstype to oppdaterteFradrag).right()
-    }
-}
-
-private fun utledPerFradragstypeOgTilhørende(
-    originaltFradrag: Fradragsgrunnlag,
-    eksterntRegulerteBeløp: EksterntRegulerteBeløp,
-): Either<ÅrsakRevurdering.BeløperMedDiff, Pair<Reguleringstype, Fradragsgrunnlag>> {
-    val fradragstype = originaltFradrag.fradragstype
-    val fradragTilhører = originaltFradrag.fradrag.tilhører
-
-    val fradragHarIkkeGrunnbeløp = !fradragstype.måJusteresVedGEndring
-    val erUtenlandskInntekt = originaltFradrag.utenlandskInntekt != null
-    if (fradragHarIkkeGrunnbeløp || erUtenlandskInntekt) {
-        return (Reguleringstype.AUTOMATISK to originaltFradrag).right()
-    }
-    if (!fradragstype.kanJusteresAutomatisk) {
-        return (
-            Reguleringstype.MANUELL(
-                ÅrsakTilManuellRegulering.ManglerRegulertBeløpForFradrag(
-                    fradragskategori = fradragstype.kategori,
-                    fradragTilhører = fradragTilhører,
-                ),
-            ) to originaltFradrag
-            ).right()
-    }
-
-    if (fradragstype.kategori == Fradragstype.Kategori.SupplerendeStønad) {
-        // Hvis eps har supplerende stønad vil samlet fradrag for eps alltid være under fribeløp og vi trenger ikke håndtere det manuelt
-        return (Reguleringstype.AUTOMATISK to originaltFradrag).right()
-    }
-
-    val eksterntBeløp = when (fradragTilhører) {
-        FradragTilhører.BRUKER -> eksterntRegulerteBeløp.beløpBruker.finn(fradragstype)
-        FradragTilhører.EPS -> eksterntRegulerteBeløp.beløpEps.finn(fradragstype)
-    }
-
-    måRevurderePåGrunnAvDifferanseMedEksterneBeløp(eksterntBeløp, originaltFradrag)?.let {
-        return it.left()
-    }
-
-    return (Reguleringstype.AUTOMATISK to originaltFradrag.oppdaterBeløpMedEksternRegulering(eksterntBeløp.etterRegulering)).right()
-}
-
-/**
- * Hvilken beløp fra Pesys som er hentet vil være basert på den samme listen med fradrag som mottas i disse metodene
- * (se [ReguleringerFraPesysService]). Det vil derfor ikke forekomme avvik
- **/
-private fun List<RegulertBeløp>.finn(fradragstype: Fradragstype) =
-    singleOrNull { it.fradragstype == EksterntBeløpSomFradragstype.from(fradragstype) }
-        ?: throw IllegalStateException("Fant ingen fradragstype $fradragstype for bruker")
-
-/**
- * Utleder reguleringstype basert på sammenligning av våre beløp og eksterne beløp.
- * Sjekker differanse før og etter regulering mot aksepterte grenser.
- *
- * @param eksterntBeløp Regulert beløp fra eksternt system
- * @param originaltFradrag Eksisterende fradragsgrunnlag
- * @return ÅrsakRevurdering.BeløperMedDiff.Fradrag eller null
- */
-private fun måRevurderePåGrunnAvDifferanseMedEksterneBeløp(
-    eksterntBeløp: RegulertBeløp,
-    originaltFradrag: Fradragsgrunnlag,
-): ÅrsakRevurdering.BeløperMedDiff.Fradrag? {
-    val vårtBeløpFørRegulering = BigDecimal(originaltFradrag.fradrag.månedsbeløp).setScale(2)
-    val eksterntBeløpFørRegulering = eksterntBeløp.førRegulering
-    val diffFørRegulering = (eksterntBeløpFørRegulering - vårtBeløpFørRegulering).abs()
-
-    if (diffFørRegulering > BigDecimal.ZERO) {
-        return ÅrsakRevurdering.BeløperMedDiff.Fradrag(
-            fradragstype = originaltFradrag.fradragstype,
-            tilhører = originaltFradrag.tilhører,
-            eksisterendeBeløp = vårtBeløpFørRegulering,
-            nyttBeløp = eksterntBeløpFørRegulering,
-        )
-    }
-
-    return null
-}
